<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
适配 scale 等不在 activation.h 中的激活算子

|config |dt|paddle old|paddle new  |pro  |pytorch  |diff  | 
|---|---|---|---|---|---|---|
|[16,16,16] |fp16|1.3510us |1.3910us | -3.0% | | % | 
|[16, 16, 1024]|fp16|2.0410us|1.7900us|14.0% | |% | 
|[16, 1024, 1024]|fp16|99.747us |84.394us| 18.2% | |% | 
|[16,16,16] |fp32|1.3930us|1.5090us | -8.0% | | % | 
|[16, 16, 1024]|fp32|2.3270us | 1.9790us| 17.6% | |% | 
|[16, 1024, 1024]|fp32|179.60us|163.88us|9.6% | |% |

因为第一个配置规模较小，因此性能会出现较小的下降，但是当规模较大时，性能会有比较明显的提升。